separate gift data and player data classes into their own packages
